+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.2-dev
 - dependency roll
+- Added `iterator()` API that returns a `DictionaryIterator`
+- Added `toArray()` API that returns an array of `HashMap` values.
 
 ## 1.0.1
 - Defined supported Node Versions. Currently Node 10 or higher.